allow the remote option to serve as fallback in conjunction with other options.

+++ src/Commands/ImportDump.php
@@ -29,7 +29,7 @@ class ImportDump extends Command
                 {--allow-production : Enable importing SQL dumps on a production system. }
                 {--force : Forces the import of the given file or remote download. Requires the dump, file or remote option. }
                 {--i|ignore-connection-filter : Ignores filter of dumps to defined connections. }
-                {--r|remote : Pull a fresh dump from the remote server as configured in the .env file. }
+                {--r|remote : Pull a fresh dump from the remote server as configured in the .env file. Will be used as fallback when combined with other options. }
                 {--flush : Delete all existing dumps in the dump folder when using a remote dump. }
                 {--l|latest : Import the most recent dump available in the configured dumps directory. }';
 
@@ -74,6 +74,7 @@ class ImportDump extends Command
         $basePath                = config('protector.baseDirectory');
         $destinationFilename     = $optionDump ?: $protector->createFilename();
         $relativePath            = $optionFile ?: $basePath . DIRECTORY_SEPARATOR . $destinationFilename;
+        $importFilePath          = null;
 
         $connectionName = null;
 
@@ -95,17 +96,40 @@ class ImportDump extends Command
         }
 
         if (!($optionRemote || $optionFile || $optionDump || $optionLatest)) {
+            $downloadRemoteDump = 'Download remote dump';
             if ($this->choice('Do you want to download and import a fresh dump from the server or an existing local dump?',
                     [
-                        '1' => 'Download remote dump',
+                        '1' => $downloadRemoteDump,
                         '2' => 'Import existing local dump',
                     ],
-                    'Download remote dump') == 'Download remote dump') {
+                    $downloadRemoteDump) == $downloadRemoteDump) {
                 $optionRemote = true;
             };
         }
 
-        if ($optionRemote) {
+        if ($optionLatest) {
+            try {
+                $importFilePath = $protector->getLatestDumpName();
+            } catch (FileNotFoundException $fileNotFoundException) {
+                if (!$optionRemote) {
+                    $this->error($fileNotFoundException->getMessage());
+                    return;
+                } else {
+                    $this->warn(sprintf('There are no files in %s', $disk->path($basePath)));
+                }
+            }
+        } elseif ($optionFile || $optionDump) {
+            if ($disk->exists($relativePath)) {
+                $importFilePath = $relativePath;
+            } elseif (!$optionRemote) {
+                $this->error((new FileNotFoundException($relativePath))->getMessage());
+                return;
+            } else {
+                $this->warn(sprintf('File not found: %s', $relativePath));
+            }
+        }
+
+        if (!$importFilePath && $optionRemote) {
             if ($this->option('flush')) {
                 $disk->delete($disk->files($basePath));
                 $this->warn(sprintf('Deleted all files in %s', $disk->path($basePath)));
@@ -123,16 +147,9 @@ class ImportDump extends Command
             $this->line(sprintf('>>> Successfully retrieved remote dump from %s', app('protector')->getServerUrl()));
 
             $importFilePath = $fullRemoteDumpFileName;
-        } elseif ($optionFile || $optionDump) {
-            $importFilePath = $relativePath;
-        } elseif ($optionLatest) {
-            try {
-                $importFilePath = $protector->getLatestDumpName();
-            } catch (FileNotFoundException $fileNotFoundException) {
-                $this->error($fileNotFoundException->getMessage());
-                return;
-            }
-        } else {
+        }
+
+        if (!$importFilePath) {
             $directoryFiles = $disk->files($basePath);
             $matchingFiles  = collect();
